Refactor garden actions, reducers, and state.
